+++ R/mod_common_forms.R
@@ -132,8 +132,16 @@ mod_common_forms_server <- function(
         )) |> 
         adjust_colnames("^SAE ")
       if(!input$show_all_data) SAE_data$subject_id <- NULL
-      datatable_custom(SAE_data, rename_vars = table_names, rownames= FALSE,
-                       title = "Serious Adverse Events", escape = FALSE)
+      
+      datatable_custom(
+        SAE_data, rename_vars = table_names, rownames= FALSE,
+        title = "Serious Adverse Events", escape = FALSE,
+        export_label = paste(
+          "SAE", 
+          ifelse(input$show_all_data, "all_patients", r$subject_id), 
+          sep = "."
+          )
+        ) 
     })
     
     output[["common_form_table"]] <- DT::renderDT({
@@ -145,8 +153,19 @@ mod_common_forms_server <- function(
           dplyr::select(-dplyr::starts_with("SAE"))
       }
       if(!input$show_all_data) df$subject_id <- NULL
-      datatable_custom(df, rename_vars = table_names, rownames= FALSE,
-                       title = form, escape = FALSE)
+      
+      datatable_custom(
+        df, 
+        rename_vars = table_names, 
+        rownames= FALSE, 
+        title = form, 
+        escape = FALSE, 
+        export_label = paste(
+          simplify_string(form), 
+          ifelse(input$show_all_data, "all_patients", r$subject_id), 
+          sep = "."
+          )
+      )
     })
     
   })