+++ R/mod_queries.R
@@ -149,11 +149,17 @@ mod_queries_server <- function(id, r, navinfo, all_forms, db_path, table_names){
         query_cols <- c("resolved", query_cols)
         table_title <- "All queries"
       }
+      
+      # determine DT dom / exts / opts
+      
+              
       datatable_custom(
         initial_queries()[query_cols], 
         table_names, 
         title = table_title,
-        callback = dblclick_to_form(ns("go_to_form"))
+        callback = dblclick_to_form(ns("go_to_form")),
+        export_label = paste(ifelse(input$show_resolved, "all", "open"),
+                           "queries", sep = ".")
       )
     })
     
@@ -184,7 +190,8 @@ mod_queries_server <- function(id, r, navinfo, all_forms, db_path, table_names){
         ),
         class = "row-border hover",
         rownames = FALSE,
-        selection = "none"
+        selection = "none",
+        allow_listing_download = FALSE
       )
     })
   })