+++ NEWS.md
@@ -13,6 +13,7 @@
 - Improved generalization of event names for use in multiple studies.
 - Make query handling a configurable option
 - Changed the legend to display 'significance pending' instead of 'significance unknown'.
+- Added `Excel` download button to Queries table & patient listings that need review.
 
 ## Bug fixes